Refactor: simplify Rust canister
