+++ canisters/ic_eth/ic_eth.did
@@ -1,5 +1,5 @@
 service : {
   verify_ecdsa : (eth_address : text, message : text, signature : text) -> (bool) query;
-  erc721_owner_of : (network : text, nft_contract_address : text, token_id : nat64) -> (text);
-  erc1155_balance_of : (network : text, nft_contract_address : text, owner_address : text, token_id : nat64) -> (nat64);
+  erc721_owner_of : (network : text, contract_address : text, token_id : nat64) -> (text);
+  erc1155_balance_of : (network : text, contract_address : text, owner_address : text, token_id : nat64) -> (nat64);
 }
